Move generateCorrelationID to top level main.go
Since the correlationID is needed by both the gitaly client and
elastic client it is best to generate above those and pass it through.
This helps as well if we end up generating a random correlation ID we'll
ensure that this random correlation ID is at least the same for both
clients.

We also remove an error branch here and just return empty string instead
of failing like the previous code since we don't really think this will
happen but even if it did we shouldn't crash the process as the
correlation ID is not necessary for the program to keep operating.