ci: Fix usage of out-of-date Gitaly version
Our CI jobs are using the Gitaly service, where we pull in the "latest"
tag. As it turns out though, "latest" is pointing to Gitaly v13.3.0-rc5,
which is definitely not the latest version at this point in time. This
is because CNG was converted to not use the "latest" tag anymore, but
instead to use a tag called "master" in gitlab-org/build/CNG!519.

Fix this by using the new "master" tag instead.